LPS-57869 renames files in message-boards-web module
