+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/messaging/ExpireBanMessageListener.java
@@ -16,7 +16,7 @@ package com.liferay.message.boards.web.messaging;
 
 import aQute.bnd.annotation.metatype.Configurable;
 
-import com.liferay.message.boards.configuration.MessageBoardsConfiguration;
+import com.liferay.message.boards.configuration.MBConfiguration;
 import com.liferay.message.boards.web.constants.MBPortletKeys;
 import com.liferay.portal.kernel.messaging.BaseSchedulerEntryMessageListener;
 import com.liferay.portal.kernel.messaging.Message;
@@ -53,7 +53,7 @@ public class ExpireBanMessageListener
 		schedulerEntry.setTriggerType(TriggerType.SIMPLE);
 
 		_messageBoardsConfiguration = Configurable.createConfigurable(
-			MessageBoardsConfiguration.class, properties);
+			MBConfiguration.class, properties);
 
 		schedulerEntry.setTriggerValue(
 			_messageBoardsConfiguration.expireBanJobInterval());
@@ -76,6 +76,6 @@ public class ExpireBanMessageListener
 	protected void setPortlet(Portlet portlet) {
 	}
 
-	private volatile MessageBoardsConfiguration _messageBoardsConfiguration;
+	private volatile MBConfiguration _messageBoardsConfiguration;
 
 }
