+++ modules/apps/message-boards/message-boards-web/src/com/liferay/message/boards/web/notifications/MBUpdateEntryUserNotificationDefinition.java
@@ -29,10 +29,10 @@ import org.osgi.service.component.annotations.Component;
 	property = {"javax.portlet.name=" + MBPortletKeys.MESSAGE_BOARDS},
 	service = UserNotificationDefinition.class
 )
-public class MessageBoardsUpdateEntryUserNotificationDefinition
+public class MBUpdateEntryUserNotificationDefinition
 	extends UserNotificationDefinition {
 
-	public MessageBoardsUpdateEntryUserNotificationDefinition() {
+	public MBUpdateEntryUserNotificationDefinition() {
 		super(
 			MBPortletKeys.MESSAGE_BOARDS, 0, NOTIFICATION_TYPE_UPDATE_ENTRY,
 			"receive-a-notification-when-someone-updates-a-post-you-are-" +
