+++ modules/apps/message-boards/message-boards-web/src/content/Language.properties
@@ -21,8 +21,8 @@ go-to-category=Go to Category
 html-format=HTML Format
 javax.portlet.description.162=Message Boards offer a conversation forum for users with multiple levels of depth and hierarchy. Administrators can manage message boards, assign user roles and permissions, and ban users. Users can manage their subscriptions to message board posts and categories, and view statistics.
 javax.portlet.long-title.162=Message Boards
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Message Boards
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Message Boards
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Message Boards
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Message Boards
 lock-thread=Lock Thread
 mailing-list=Mailing List
 mark-as-a-question=Mark as a Question
