+++ modules/apps/message-boards/message-boards-web/src/content/Language_ar.properties
@@ -21,8 +21,8 @@ go-to-category=الذهاب إلى التصنيف
 html-format=تنسيق HTML
 javax.portlet.description.162=المنتديات توفر منتدى المحادثة للمستخدمين مع مستويات متعددة من العمق والتسلسل الهرمي. يمكن لمسؤولي إدارة لوحات الرسائل، تعيين الأدوار للمستخدمين والأذونات، وحظر المستخدمين . يمكن للمستخدمين إدارة اشتراكاتهم في المنديات والفئات، وعرض الإحصاءات.
 javax.portlet.long-title.162=المنتديات (لوحة التحكم)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=المنتديات
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=المنتديات
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=المنتديات
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=المنتديات
 lock-thread=قفل المشاركة
 mailing-list=القائمة البريدية
 mark-as-a-question=وضع علامة سؤال
