+++ modules/apps/message-boards/message-boards-web/src/content/Language_ca.properties
@@ -21,8 +21,8 @@ go-to-category=Ves a la categoria
 html-format=Format HTML
 javax.portlet.description.162=Els fòrums de discussió són un lloc de conversa per als usuaris amb diferents nivells de profunditat i jerarquia. Els administradors poden gestionar els fòrums, assignar rols i permisos d'usuari i expulsar usuaris. Per contra, els usuaris poden gestionar les subscripcions que facin als apunts i a les categories del fòrum, així com veure estadístiques.
 javax.portlet.long-title.162=Fòrums de discussió
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Fòrums de discussió
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Fòrums de discussió
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Fòrums de discussió
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Fòrums de discussió
 lock-thread=Bloqueja el fil de la discussió
 mailing-list=Llista de correu
 mark-as-a-question=Marca com a pregunta
