+++ modules/apps/message-boards/message-boards-web/src/content/Language_cs.properties
@@ -21,8 +21,8 @@ go-to-category=Přejít do kategorie
 html-format=HTML formát
 javax.portlet.description.162=Diskusní fórum je místo, kde je možné vést konverzace mnoha uživatelů na různá témata, podle kterých je možné diskuse hierarchicky rozčlenit. Správci pak mohou diskuse zakládat, spravovat a přesouvat, mohou také přiřazovat uživatelské role a oprávnění, nebo zamezovat různým uživatelům přístup. Běžný uživatel má také možnost přihlásit se k odběru novinek (sledovat) z různých kategorií a tím být neustále e-mailem informován o nových příspěvcích.
 javax.portlet.long-title.162=Diskusní fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Diskusní fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Diskusní fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Diskusní fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Diskusní fórum
 lock-thread=Zamknout diskusi
 mailing-list=Mailing-list
 mark-as-a-question=Označit jako otázku
