+++ modules/apps/message-boards/message-boards-web/src/content/Language_da.properties
@@ -21,8 +21,8 @@ go-to-category=Gå til kategori
 html-format=HTML Format (Automatic Copy)
 javax.portlet.description.162=Message Boards offer a conversation forum for users with multiple levels of depth and hierarchy. Administrators can manage message boards, assign user roles and permissions, and ban users. Brugere can manage their subscriptions to message board posts and categories, and view statistics.
 javax.portlet.long-title.162=Forum (Kontrolpanel)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Diskussionsfora
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Diskussionsfora
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Diskussionsfora
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Diskussionsfora
 lock-thread=Lås tråd
 mailing-list=E-mail-liste
 mark-as-a-question=Mark as a Question (Automatic Copy)
