+++ modules/apps/message-boards/message-boards-web/src/content/Language_de.properties
@@ -21,8 +21,8 @@ go-to-category=Zu Kategorie wechseln
 html-format=HTML-Format
 javax.portlet.description.162=Foren bieten eine vielschichtige Gesprächs- und Austauschplattform für Benutzer. Administratoren können Foren verwalten, Benutzerrollen und Berechtigungen zuweisen und Nutzer verbannen. Benutzer können ihre Abonnements zu Forenbeiträgen oder Kategorien verwalten und Foren-Statistiken einsehen.
 javax.portlet.long-title.162=Foren
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Foren
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Foren
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Foren
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Foren
 lock-thread=Thread sperren
 mailing-list=Mailingliste
 mark-as-a-question=Als Frage markieren
