+++ modules/apps/message-boards/message-boards-web/src/content/Language_el.properties
@@ -21,8 +21,8 @@ go-to-category=Μετάβαση στην κατηγορία
 html-format=Μορφή HTML
 javax.portlet.description.162=Οι Πίνακες Μηνυμάτων προσφέρουν ένα φόρουμ συνομιλίας για τους χρήστες, με πολλαπλά επίπεδα βάθους και ιεραρχίας. Οι διαχειριστές μπορούν να επεξεργαστούν τους πίνακες μηνυμάτων, να ορίσουν τους ρόλους και δικαιώματα χρηστών για αυτά και να κλειδώσουν χρήστες εκτός αυτών. Οι χρήστες μπορούν να χειρίζοτναι τις συνδρομές τους στα άρθρα και τις κατηγορίες πινάκων μηνυμάτων, και να δουν σχετικά στατιστικά στοιχεία.
 javax.portlet.long-title.162=Πίνακες Μηνυμάτων (πίνακας ελέγχου)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Πίνακες Μηνυμάτων
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Πίνακες Μηνυμάτων
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Πίνακες Μηνυμάτων
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Πίνακες Μηνυμάτων
 lock-thread=Κλείδωμα thread
 mailing-list=Κατάλογος διευθύνσεων
 mark-as-a-question=Σήμανση ως ερώτηση
