+++ modules/apps/message-boards/message-boards-web/src/content/Language_es.properties
@@ -21,8 +21,8 @@ go-to-category=Ir a la categoría
 html-format=Formato HTML
 javax.portlet.description.162=Los foros son un lugar de conversación para los usuarios con múltiples niveles de profundidad y jerarquía. Los administradores pueden gestionar los foros, asignar roles y permisos de usuario, y expulsar usuarios. Los usuarios pueden gestionar sus suscripciones a los mensajes y categorías del foro, así como ver las estadísticas.
 javax.portlet.long-title.162=Foros de Discusión
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Foro de discusión
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Foros
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Foro de discusión
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Foros
 lock-thread=Bloquear hilo de discusión
 mailing-list=Lista de correo
 mark-as-a-question=Marcar como pregunta
