+++ modules/apps/message-boards/message-boards-web/src/content/Language_et.properties
@@ -21,8 +21,8 @@ go-to-category=Mine kategooria juurde
 html-format=HTML formaat
 javax.portlet.description.162=Message Boards offer a conversation forum for users with multiple levels of depth and hierarchy. Administrators can manage message boards, assign user roles and permissions, and ban users. Users can manage their subscriptions to message board posts and categories, and view statistics. (Automatic Copy)
 javax.portlet.long-title.162=Foorumid
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Foorumid
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Foorumid
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Foorumid
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Foorumid
 lock-thread=Lock Thread (Automatic Copy)
 mailing-list=Postiloend
 mark-as-a-question=Märgi küsimusena
