+++ modules/apps/message-boards/message-boards-web/src/content/Language_eu.properties
@@ -21,8 +21,8 @@ go-to-category=Kategoriara joan
 html-format=HTML formatua
 javax.portlet.description.162=Iragarki Taulak, erabiltzaileei hainbat sakonera-maila eta hierarkiako eztabaidagune bat eskaintzen dio. Administratzaileek iragarki taulak kudeatu, erabiltzaileei rol eta baimenak esleitu, eta erabiltzaileak debekatu ditzakete. Erabiltzaileek iragarki tauletako beraien mezu eta kategori harpidetzak kudeatu, eta estatistikak ikusi ditzekete.
 javax.portlet.long-title.162=Foroa (Kontrol Panela)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Iragarki taula
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Iragarki Taulak
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Iragarki taula
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Iragarki Taulak
 lock-thread=Blokeatu haria
 mailing-list=Posta zerrenda
 mark-as-a-question=Markatu galdera gisa
