+++ modules/apps/message-boards/message-boards-web/src/content/Language_fa.properties
@@ -21,8 +21,8 @@ go-to-category=برو به گروهبندی
 html-format=قالب HTML
 javax.portlet.description.162=تابلو پیام محلی برای تعامل بین کاربران است. در این بخش امکان گروه‌بندی پیام‌ها در چندین سطح وجود دارد. کاربر مدیر می‌تواند مجوز و نقش‌های مربوط به این قسمت را در اختیار کاربران قرار دهد یا کاربران را به عنوان کاربران غیرمجاز تعریف کرده و به این ترتیب دسترسی آنها را به بخش تابلو پیام مسدود کند. کاربران نیز قادر هستند عضویت خود را در پیام‌های موجود اعلام نمایند. به این ترتیب پیام‌های درج شده به آدرس ایمیل ایشان ارسال می‌شود. همچنین کاربران قادر به مشاهده و دسته‌بندی پیام‌های خود هستند.
 javax.portlet.long-title.162=تالار گفتگو
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=تالار گفتگو
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=تالار گفتگو
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=تالار گفتگو
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=تالار گفتگو
 lock-thread=قفل کردن موضوع
 mailing-list=لیست ایمیل‌ها
 mark-as-a-question=علامت زدن به عنوان سوال
