+++ modules/apps/message-boards/message-boards-web/src/content/Language_fi.properties
@@ -21,8 +21,8 @@ go-to-category=Siirry luokitteluun
 html-format=HTML-muoto
 javax.portlet.description.162=Keskustelualueet tarjoavat useatasoisen keskustelukanavan käyttäjille. Ylläpitäjät voivat hallita keskustelualueita ja määrittää käyttäjille rooleja ja oikeuksia ja antaa porttikieltoja. Käyttäjät voivat hallita keskustelujaan ja aihealueitaan, sekä katsoa tilastoja.
 javax.portlet.long-title.162=Keskusteluryhmät
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Keskustelualueet
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Keskustelualueet
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Keskustelualueet
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Keskustelualueet
 lock-thread=Lukitse keskustelu polku
 mailing-list=Sähköpostituslista
 mark-as-a-question=Merkkaa kysymykseksi
