+++ modules/apps/message-boards/message-boards-web/src/content/Language_fr.properties
@@ -21,8 +21,8 @@ go-to-category=Aller à la catégorie
 html-format=Format de HTML
 javax.portlet.description.162=Les forums offrent un espace de discussion pour les utilisateurs, avec de multiples niveaux de profondeur et une hiérarchie. Les administrateurs peuvent gérer les forums, assigner rôles d'utilisateur et permissions, et bannir les utilisateurs. Les utilisateurs peuvent gérer leurs abonnements aux messages et catégories, et voir les statistiques.
 javax.portlet.long-title.162=Forums (Panneau de configuration)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Forum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Forum
 lock-thread=Fermez à clef le fil
 mailing-list=Liste d'adresses
 mark-as-a-question=Marque comme question
