+++ modules/apps/message-boards/message-boards-web/src/content/Language_gl.properties
@@ -21,8 +21,8 @@ go-to-category=Ir á categoría
 html-format=Formato HTML
 javax.portlet.description.162=Os foros de discusión ofrecen un lugar de converesa con multiples niveis de xerarquía e profundidade. Os administradores poden administrar os foros, asignar roles de usuario, permisos e expulsar usuarios. Os usuarios poden administrar as súas subscricións nos temas e categorías do foro, e ver estatísticas.
 javax.portlet.long-title.162=Foro de discusión
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Foro de discusión
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Foros de discusión
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Foro de discusión
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Foros de discusión
 lock-thread=Bloquear fío
 mailing-list=Lista de usuarios a quen se envía propaganda
 mark-as-a-question=Marcar como pregunta
