+++ modules/apps/message-boards/message-boards-web/src/content/Language_hi_IN.properties
@@ -21,8 +21,8 @@ go-to-category=श्रेणी पर जाएँ
 html-format=HTML प्रारूप
 javax.portlet.description.162=संदेश बोर्डों गहराई और पदानुक्रम के अनेक स्तरों के साथ उपयोगकर्ताओं के लिए एक बातचीत के मंच की पेशकश करते हैं। व्यवस्थापक संदेश बोर्डों का प्रबंधन करें, उपयोगकर्ता भूमिकाएँ और अनुमतियाँ असाइन कर सकते हैं, और उपयोगकर्ता पर प्रतिबंध। उपयोगकर्ताओं को संदेश बोर्ड पदों और श्रेणियाँ, और आँकड़े देखने के लिए अपनी सदस्यता का प्रबंधन कर सकते हैं। (Automatic Translation)
 javax.portlet.long-title.162=संदेश बोर्ड्स
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=संदेश बोर्ड्स
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=संदेश बोर्ड्स
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=संदेश बोर्ड्स
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=संदेश बोर्ड्स
 lock-thread=थ्रेड को लॉक करना
 mailing-list=मेलिंग सूची
 mark-as-a-question=एक प्रश्न के रूप में चिह्नित करें (Automatic Translation)
