+++ modules/apps/message-boards/message-boards-web/src/content/Language_hr.properties
@@ -21,8 +21,8 @@ go-to-category=Idi na kategoriju
 html-format=HTML oblik
 javax.portlet.description.162=Oglasne ploče pružaju korisnicima forum za rasprave s višerazinskom dubinom i hijerarhijom. Administratori mogu upravljati oglasnim pločama, dodjeljivati korisničke uloge i dopuštenja i zabraniti pristup korisnicima. Korisnici mogu upravljati svojim pretplatama na objave sa oglasnih ploča i kategorije, te pregledavati statistike.
 javax.portlet.long-title.162=Oglasne ploče
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Oglasne ploče
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Oglasne ploče
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Oglasne ploče
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Oglasne ploče
 lock-thread=Lock teme
 mailing-list=Mailing lista
 mark-as-a-question=Označi kao pitanje
