+++ modules/apps/message-boards/message-boards-web/src/content/Language_hu.properties
@@ -21,8 +21,8 @@ go-to-category=Kategóriához lépés
 html-format=HTML formázás
 javax.portlet.description.162=A fórum hierarchikusan szervezhető csatornákon keresztül biztosítja a felhasználók közötti eszmecserét. Az adminisztrátorok kezelik, ők rendelhetik hozzá a felhasználói szerepköröket és engedélyeket, valamint ők tilthatják le a felhasználókat. A felhasználók kezelhetik, milyen hozzászólásokra és kategóriákra iratkoznak fel, és megtekinthetik a statisztikákat.
 javax.portlet.long-title.162=Fórumok
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Fórum
 lock-thread=Szál lezárása
 mailing-list=Levelező lista
 mark-as-a-question=Megjelölés kérdésként
