+++ modules/apps/message-boards/message-boards-web/src/content/Language_in.properties
@@ -21,8 +21,8 @@ go-to-category=Pergi ke Kategori
 html-format=Format HTML
 javax.portlet.description.162=Message Boards menawarkan forum percakapan untuk pengguna dengan berbagai tingkat kategori dan hirarki. Administrator dapat mengatur message bard, memberikan peran pengguna dan hak akses, dan memblokir pengguna tertentu. Pengguna dapat mengelola subscription untuk posting message board dan kategori tertentu, dan melihat statistik
 javax.portlet.long-title.162=Pesan Papan (Control Panel)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Message Board
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Message Board
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Message Board
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Message Board
 lock-thread=Thread Kunci
 mailing-list=Daftar Mailing
 mark-as-a-question=Tandai sebagai Pertanyaan
