+++ modules/apps/message-boards/message-boards-web/src/content/Language_it.properties
@@ -21,8 +21,8 @@ go-to-category=Vai alla Categoria
 html-format=Formato HTML
 javax.portlet.description.162=I Forum offrono agli utenti uno spazio di conversazione con più livelli di profondità e gerarchia. Gli amministratori possono gestire i forum, assegnare i ruoli utente, le autorizzazioni e bloccare gli utenti. Gli utenti possono gestire le proprie iscrizioni alle discussioni, alle categorie e visualizzare le statistiche.
 javax.portlet.long-title.162=Forum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Forum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Forum
 lock-thread=Blocca Thread
 mailing-list=Mailing List
 mark-as-a-question=Segna come Domanda
