+++ modules/apps/message-boards/message-boards-web/src/content/Language_iw.properties
@@ -21,8 +21,8 @@ go-to-category=לך לקטגוריה
 html-format=פורמט HTML
 javax.portlet.description.162=לוחות הודעות מציעים פורום שיחות עבור משתמשים ברמות עומק ומדרג רבות. מנהלים יכולים לנהל לוחות הודעות, להקצות תפקידים ואישורים למשתמשים ולחסום משתמשים. משתמשים יכולים לנהל את המנויים שלהם לפוסטים ולקטגוריות של לוח ההודעות, ולצפות בנתונים סטטיסטיים.
 javax.portlet.long-title.162=Message Boards (Control Panel)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=לוחות מודעות
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=לוחות מודעות
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=לוחות מודעות
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=לוחות מודעות
 lock-thread=נעילת אשכול
 mailing-list=רשימת דיוור
 mark-as-a-question=סמן בתור שאלה
