+++ modules/apps/message-boards/message-boards-web/src/content/Language_ja.properties
@@ -21,8 +21,8 @@ go-to-category=カテゴリを参照する
 html-format=HTMLのフォーマット
 javax.portlet.description.162=掲示板では、ユーザーは様々なユーザーと様々なレベルでやりとりができます。管理者は、ユーザーロールや権限の割り当てや、ユーザーを掲示板にアクセス禁止することができます。ユーザーは投稿やカテゴリの変更通知を管理、掲示板の統計を見ることもできます。
 javax.portlet.long-title.162=掲示板
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=掲示板
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=掲示板
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=掲示板
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=掲示板
 lock-thread=スレッドをロック
 mailing-list=メーリングリスト
 mark-as-a-question=質問として設定する
