+++ modules/apps/message-boards/message-boards-web/src/content/Language_ko.properties
@@ -21,8 +21,8 @@ go-to-category=종류에 가십시오
 html-format=HTML체재
 javax.portlet.description.162=메시지 보드 사용자는 깊이 계층의 여러 수준에 대 한 대화 포럼을 제공합니다. 관리자, 메시지 보드를 관리할 수 사용자 역할 및 권한, 할당 하 고 사용자를 금지. 보드 게시물 및 범주, 메시지 및 통계를 볼에 대 한 구독을 관리할 수 있습니다. (Automatic Translation)
 javax.portlet.long-title.162=메시지는 난입한다
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=메시지는 난입한다
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=메시지는 난입한다
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=메시지는 난입한다
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=메시지는 난입한다
 lock-thread=잠금 (Automatic Translation)
 mailing-list=우편 목록
 mark-as-a-question=질문으로 표
