+++ modules/apps/message-boards/message-boards-web/src/content/Language_lo.properties
@@ -21,8 +21,8 @@ go-to-category=ໄປຍັງໝວດໝູ່
 html-format=ຮູບແບບ HTML
 javax.portlet.description.162=ກະດານຂໍ້ຄວາມຕ້ອງການຫ້ອງການສົນທະນາສຳລັບຜູ້ໃຊ້ກັບຫຼາຍໆລະດັບ ແລະຂັ້ນທີ່ຕ່າງກັນ. ຜູ້ບໍລິຫານສາມາດບໍລິຫານຄຸ້ມຄອງກະດານຂໍ້ຄວາມ, ກຳນົດຫຼັກການ ແລະສິດຕ່າງໆ ໃຫ້ກັບຜູ້ໃຊ້ ແລະລົງໂທດຜູ້ໃຊ້. ຜູ້ໃຊ້ສາມາດບໍລິຫານຊັບສກຣິບເຊີ້ນສົ່ງໄປຍັງກະດານຂໍ້ຄວາມ , ໝວດຕ່າງໆ ແລະການສະແດງສະຖິຄິຕ່າງໆ
 javax.portlet.long-title.162=ກະດານຂໍ້ຄວາມ (ຄອນໂທນພາເນວ)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=ການດານຂໍ້ຄວາມ
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=ການດານຂໍ້ຄວາມ
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=ການດານຂໍ້ຄວາມ
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=ການດານຂໍ້ຄວາມ
 lock-thread=ລ໋ອກທິຣດ
 mailing-list=ບັນຊີເມວລິ້ງ
 mark-as-a-question=ມາກຄືກັນກັບຄຳຖາມ
