+++ modules/apps/message-boards/message-boards-web/src/content/Language_nb.properties
@@ -21,8 +21,8 @@ go-to-category=Gå til kategori
 html-format=HTML-format
 javax.portlet.description.162=Diskusjonsfora tilbyr forum med flere nivåer til brukere. Administratorer kan administrere diskusjonsfora, tilegne brukere roller og rettigheter, og utestenge brukere. Brukere kan abonnere på poster og kategorier, samt se statistikk.
 javax.portlet.long-title.162=Diskusjonsfora
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Diskusjonsfora
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Diskusjonsfora
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Diskusjonsfora
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Diskusjonsfora
 lock-thread=Lås tråd
 mailing-list=E-postliste
 mark-as-a-question=Marker som et spørsmål
