+++ modules/apps/message-boards/message-boards-web/src/content/Language_nl.properties
@@ -21,8 +21,8 @@ go-to-category=Naar categorie gaan
 html-format=HTML-indeling
 javax.portlet.description.162=Een Forum is een conversatieplatform voor gebruikers dat bestaat uit meerdere niveaus. Een beheerder kan forums beheren, gebruikersrollen en rechten toewijzen en gebruikers uitsluiten. Een gebruiker kan conversaties en categorieën volgen per e-mail, en statistieken bekijken.
 javax.portlet.long-title.162=Forums
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Forums
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Forums
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Forums
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Forums
 lock-thread=Conversatie vergrendelen
 mailing-list=Mailinglijst
 mark-as-a-question=Als vraag markeren
