+++ modules/apps/message-boards/message-boards-web/src/content/Language_pl.properties
@@ -21,8 +21,8 @@ go-to-category=Przejdź do kategorii
 html-format=Format HTML
 javax.portlet.description.162=Fora oferują użytkownikom moduł forum dyskusyjnego o rożnych poziomach głębokości i hierarchii. Administratorzy mogą zarządzać forum, przydzielać użytkownikom role i uprawnienia oraz blokować użytkowników. Użytkownicy mogą zarządzać subskrypcjami postów na forum i kategoriami oraz przeglądać statystyki.
 javax.portlet.long-title.162=Fora
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Fora dyskusyjne
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Fora
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Fora dyskusyjne
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Fora
 lock-thread=Zamknij wątek
 mailing-list=Lista mailingowa
 mark-as-a-question=Oznacz jako pytanie
