+++ modules/apps/message-boards/message-boards-web/src/content/Language_pt_BR.properties
@@ -21,8 +21,8 @@ go-to-category=Ir para categoria
 html-format=Formato HTML
 javax.portlet.description.162=Fórum é um local de discussão entre os usuários com vários níveis de categorização. Administradores podem gerenciar fóruns, controlar permissões e banir usuários. Usuários podem gerencias suas subscrições em categorias do fórum e visualizar estatísticas.
 javax.portlet.long-title.162=Fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Fórum
 lock-thread=Bloquear tópico
 mailing-list=Mailing list
 mark-as-a-question=Marcar como pergunta
