+++ modules/apps/message-boards/message-boards-web/src/content/Language_pt_PT.properties
@@ -21,8 +21,8 @@ go-to-category=Vá à categoria
 html-format=Formato do HTML
 javax.portlet.description.162=Os Fóruns oferecem um espaço de discussão para os utilizadores com vários níveis de profundidade e hierarquia. Um Administrador pode gerir os Fóruns, atribuir perfís e permissões e remover utilizadores. Um utilizador pode administrar as categorias que subscreve, assim como consultar as estatísticas.
 javax.portlet.long-title.162=Fóruns (Painel de Controlo)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Fóruns
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Fóruns
 lock-thread=Bloquear Tópico
 mailing-list=Lista de endereços
 mark-as-a-question=Marca como uma pergunta
