+++ modules/apps/message-boards/message-boards-web/src/content/Language_ro.properties
@@ -21,8 +21,8 @@ go-to-category=Mergi la Categorie
 html-format=Format HTML
 javax.portlet.description.162=Forumurile oferă utlizatorilor o modalitate de a discuta în cadrul unor nivele de adâncime și ierarhie diferite. Administatorii pot gestiona forumurile, asocia utilizatori și permisiuni, și interzice utilizatori. Utilizatorii pot gestiona subscripțiile lor la discuțiile și categoriile forumurilor, și pot vizualiza date statistice.
 javax.portlet.long-title.162=Forumuri
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Forumuri
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Forumuri
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Forumuri
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Forumuri
 lock-thread=Blocheaza Discutie
 mailing-list=Lista de destinatari
 mark-as-a-question=Marcheaza ca Intrebare
