+++ modules/apps/message-boards/message-boards-web/src/content/Language_ru.properties
@@ -21,8 +21,8 @@ go-to-category=Перейти к категории
 html-format=Формат HTML
 javax.portlet.description.162=Доски сообщений представляют собой форум для общения пользователей с иерархией и множеством уровней вложенности. Администраторы могут управлять досками сообщений, назначать пользователям роли и права доступа и помещать пользователей в черный список. Пользователи могут управлять своими подписками на сообщения и разделы досок сообщений и просматривать статистику.
 javax.portlet.long-title.162=Форумы
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Форумы
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Доски сообщений
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Форумы
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Доски сообщений
 lock-thread=Закрыть нить.
 mailing-list=Список рассылки
 mark-as-a-question=Пометить как вопрос
