+++ modules/apps/message-boards/message-boards-web/src/content/Language_sk.properties
@@ -21,8 +21,8 @@ go-to-category=Prejsť na kategóriu
 html-format=HTML formát
 javax.portlet.description.162=Diskusné fórum je miesto, kde je možné viesť konverzáciu viacero používateľov na rôzne témy, podľa ktorých je možné diskusie hierarchicky usporiadať. Administrátori môžu zakladať, spravovať, presúvať diskusie a môžu priraďovať používateľské postavenia a práva, alebo vymedzovať rôznym používateľom prístup. Bežný používateľ má tiež možnosť prihlásiť sa k odberu noviniek (sledovať) z rôznych kategórií a tým byť neustále e-mailom informovaný o nových príspevkoch.
 javax.portlet.long-title.162=Diskusné fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Diskusné fórum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Diskusné fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Diskusné fórum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Diskusné fórum
 lock-thread=Zamkni diskusiu
 mailing-list=Mailing-list
 mark-as-a-question=Označiť ako otázku
