+++ modules/apps/message-boards/message-boards-web/src/content/Language_sl.properties
@@ -21,8 +21,8 @@ go-to-category=Pojdi na kategorijo
 html-format=HTML format
 javax.portlet.description.162=Message Boards offer a conversation forum for users with multiple levels of depth and hierarchy. Administrators can manage message boards, assign user roles and permissions, and ban users. Users can manage their subscriptions to message board posts and categories, and view statistics. (Automatic Copy)
 javax.portlet.long-title.162=Table s sporočili
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Table s sporočili
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Table s sporočili
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Table s sporočili
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Table s sporočili
 lock-thread=Lock Thread (Automatic Copy)
 mailing-list=E-poštni seznam
 mark-as-a-question=Označi kot vprašanje
