+++ modules/apps/message-boards/message-boards-web/src/content/Language_sr_RS.properties
@@ -21,8 +21,8 @@ go-to-category=Иди на Категорију
 html-format=ХТМЛ Формат
 javax.portlet.description.162=Огласне Табле нуде форум за разговор корисника са више нивоа дубине и хијерархије. Администратори могу управљати форумима, доделити улоге корисника и дозволе и забране корисника. Корисници могу да управљају својим претплатама на постове огласне табле и категорије, као и преглед статистике.
 javax.portlet.long-title.162=Огласне Табле
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Форум
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Огласне Табле
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Форум
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Огласне Табле
 lock-thread=Закључај Тему
 mailing-list=Мејлинг Листа
 mark-as-a-question=Означи као Питање
