+++ modules/apps/message-boards/message-boards-web/src/content/Language_sr_RS_latin.properties
@@ -21,8 +21,8 @@ go-to-category=Idi na Kategoriju
 html-format=HTML Format (Automatic Copy)
 javax.portlet.description.162=Oglasne Table nude forum za razgovor korisnika sa više nivoa dubine i hijerarhije. Administratori mogu upravljati forumima, dodeliti uloge korisnika i dozvole i zabrane korisnika. Korisnici mogu da upravljaju svojim pretplatama na postove oglasne table i kategorije, kao i pregled statistike.
 javax.portlet.long-title.162=Message Boards (Control Panel)
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Forum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Oglasne Table
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Oglasne Table
 lock-thread=Zaključaj Temu
 mailing-list=Mejling Lista
 mark-as-a-question=Označi kao Pitanje
