+++ modules/apps/message-boards/message-boards-web/src/content/Language_sv.properties
@@ -21,8 +21,8 @@ go-to-category=Gå till kategori
 html-format=HTML-format
 javax.portlet.description.162=Forum erbjuder användare en plats där de kan konversera i en miljö med multipla nivåer och hierarkisk struktur. Administratörer kan hantera forum, tilldela användarroller och rättigheter och stänga av användare. Användare kan hantera deras prenumerationer till foruminlägg och forumtrådar och se statistik.
 javax.portlet.long-title.162=Diskussionsforum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Forum
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Forum
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Forum
 lock-thread=Lås tråd
 mailing-list=Epost-sändlista
 mark-as-a-question=Markera som fråga
