+++ modules/apps/message-boards/message-boards-web/src/content/Language_tr.properties
@@ -21,8 +21,8 @@ go-to-category=Go to Category (Automatic Copy)
 html-format=HTML Formatı
 javax.portlet.description.162=Message Boards offer a conversation forum for users with multiple levels of depth and hierarchy. Administrators can manage message boards, assign user roles and permissions, and ban users. Users can manage their subscriptions to message board posts and categories, and view statistics. (Automatic Copy)
 javax.portlet.long-title.162=Mesaj Tahtası
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Mesaj Tahtası
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Mesaj Tahtası
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Mesaj Tahtası
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Mesaj Tahtası
 lock-thread=Lock Thread (Automatic Copy)
 mailing-list=E-Posta Listesi
 mark-as-a-question=Soru Olarak İşaratle
