+++ modules/apps/message-boards/message-boards-web/src/content/Language_vi.properties
@@ -21,8 +21,8 @@ go-to-category=Đi tới Chuyên mục
 html-format=Định dạng HTML
 javax.portlet.description.162=<Diễn đàn> là sân chơi thảo luận với các chủ đề cấu trúc từ lớn đến nhỏ, nhiều mức và có tính phân cấp. Quản trị viên có quyền quản lý diễn đàn, gán vai trò và phân quyền, hoặc cấm người sử dụng vi phạm. Người sử dụng có thể đăng ký nhận thông tin với từng chủ đề và xem trạng thái của luồng thảo luận.
 javax.portlet.long-title.162=Diễn đàn
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=Diễn đàn
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=Diễn đàn
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=Diễn đàn
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=Diễn đàn
 lock-thread=Khóa chủ đề
 mailing-list=Danh sách gửi thư
 mark-as-a-question=Là câu hỏi
