+++ modules/apps/message-boards/message-boards-web/src/content/Language_zh_CN.properties
@@ -21,8 +21,8 @@ go-to-category=访问类别
 html-format=HTML格式
 javax.portlet.description.162=留言板为用户提供了一个囊括多种层级、多种深度的会话论坛。管理员可以管理留言板、赋予用户角色和权限、冻结用户等。用户可以管理自己对于留言板上帖子和类别的订阅，还能查看统计信息。
 javax.portlet.long-title.162=留言板
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=留言板
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=留言板
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=留言板
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=留言板
 lock-thread=锁定话题
 mailing-list=邮件列表
 mark-as-a-question=标记为问题
