+++ modules/apps/message-boards/message-boards-web/src/content/Language_zh_TW.properties
@@ -21,8 +21,8 @@ go-to-category=到類別
 html-format=HTML 格式
 javax.portlet.description.162=討論版為使用者提供一個含有多個層次與階層的方便論壇。系統管理員能管理討論版，指派使用者角色與權限，和禁止使用者。使用者能管理他們對帖子、類別的訂閱，並能檢視統計資料。
 javax.portlet.long-title.162=討論板
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsPortlet=討論板
-javax.portlet.title.com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet=討論板
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBPortlet=討論板
+javax.portlet.title.com_liferay_message_boards_web_portlet_MBAdminPortlet=討論板
 lock-thread=鎖定主題
 mailing-list=郵件表
 mark-as-a-question=標記為問題
