+++ modules/apps/message-boards/message-boards-web/src/resource-actions/default.xml
@@ -3,7 +3,7 @@
 
 <resource-action-mapping>
 	<portlet-resource>
-		<portlet-name>com_liferay_message_boards_web_portlet_MessageBoardsPortlet</portlet-name>
+		<portlet-name>com_liferay_message_boards_web_portlet_MBPortlet</portlet-name>
 		<permissions>
 			<supports>
 				<action-key>ADD_TO_PAGE</action-key>
@@ -22,7 +22,7 @@
 		</permissions>
 	</portlet-resource>
 	<portlet-resource>
-		<portlet-name>com_liferay_message_boards_web_portlet_MessageBoardsAdminPortlet</portlet-name>
+		<portlet-name>com_liferay_message_boards_web_portlet_MBAdminPortlet</portlet-name>
 		<permissions>
 			<supports>
 				<action-key>ACCESS_IN_CONTROL_PANEL</action-key>
